# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cẩm Phả GIS Server API
  version: 1.0.0
  description: API quản lý dữ liệu GIS và ngập lụt TP Cẩm Phả.
servers:
  - url: http://localhost:3005
    description: Local development
  - url: / 
    description: Current host
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/ldap/login:
    post:
      tags: [Auth]
      operationId: ldapLogin
      summary: Authenticate a pre-provisioned organization account through LDAPS
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LdapLoginRequest' }
      responses:
        '200':
          description: AD authentication succeeded and JWT tokens were issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503':
          description: Organization directory temporarily unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string, writeOnly: true }
      responses:
        '200': { description: Refresh token rotated, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/admin/users:
    get:
      tags: [Users]
      operationId: listUsers
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: roleCode, in: query, schema: { $ref: '#/components/schemas/RoleCode' } }
        - { name: q, in: query, description: 'Tìm theo id, email, họ tên, số điện thoại, vai trò, hoặc tên đăng nhập AD (auth.ldap_identities.login_name)', schema: { type: string, maxLength: 255 } }
      responses:
        '200':
          description: Organization-scoped users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created in actor organization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Organization-scoped user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User soft-deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/users/{id}/role:
    patch:
      tags: [Users]
      operationId: changeUserRole
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleCode]
              properties:
                roleCode: { $ref: '#/components/schemas/RoleCode' }
      responses:
        '200':
          description: Role changed; only actors with users.change_role
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
  schemas:
    RoleCode:
      type: string
      enum: [system_admin, ubnd_tp, so_tnmt, so_xd, citizen]
...
    LdapLoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 255
          pattern: '^[A-Za-z0-9][A-Za-z0-9._@-]*$'
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
...
    CreateUserRequest:
      oneOf:
        - type: object
          additionalProperties: false
          required: [email, password, fullName]
          properties:
            authProvider: { type: string, const: local, default: local }
            email: { type: string, format: email }
            password: { type: string, minLength: 8, maxLength: 128, writeOnly: true }
            fullName: { type: string, minLength: 2, maxLength: 255 }
            phone: { type: [string, 'null'] }
            roleCode: { $ref: '#/components/schemas/RoleCode' }
        - type: object
          additionalProperties: false
          required: [authProvider, directoryUsername]
          properties:
            authProvider: { type: string, const: ldap }
            directoryUsername:
              type: string
              maxLength: 255
              pattern: '^[A-Za-z0-9][A-Za-z0-9._@-]*$'
            roleCode: { $ref: '#/components/schemas/RoleCode' }
...
    SuccessResponse:
      type: object
      required: [message, status, data]
      properties:
        message: { type: string }
        status: { type: integer }
        data: {}
        metadata: { type: object }
    ListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              required: [items]
              properties:
                items: { type: array, items: {} }
            metadata:
              type: object
              required: [page, limit, total, totalPages]
              properties:
                page: { type: integer }
                limit: { type: integer }
                total: { type: integer }
                totalPages: { type: integer }
    ErrorResponse:
      type: object
      required: [success, message, errors]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        errors:
          type: array
          items: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Role permission or organization scope denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }